tạo các exception bắt lỗi, thêm các DTO cho meeting_, view sent request, edit profile, view , view booked slot

+++ Meet-My-Lecture/src/main/java/com/springboot/meetMyLecturer/controller/LecturerController.java
@@ -61,7 +61,7 @@ public class LecturerController {
 
     // SLOT : CREATE EMPTY SLOT
     @PostMapping("slots/{userId}/users")
-    public ResponseEntity<EmptySlotDTO> createEmptySlot(@PathVariable(value = "userId") int userId,
+    public ResponseEntity<EmptySlotDTO> createEmptySlot(@PathVariable(value = "userId") Long userId,
                                                         @RequestBody EmptySlot emptySlot) {
 
         EmptySlotDTO emptySlotDTO = emptySlotService.creatEmptySlot(userId, emptySlot);
