+++ Meet-My-Lecture/src/main/java/com/springboot/meetMyLecturer/repository/UserRepository.java
@@ -8,12 +8,12 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
     @Query( "select u from User u join Role r on u.role.roleId = r.roleId where u.userName like %:name% and r.roleName ='LECTURER' ")
     List<User> findLecturerByUserName(String name);
 
-    User findUserByUserId(int id);
+    User findUserByUserId(long id);
 
 
 }
