+++ Meet-My-Lecture/src/main/java/com/springboot/meetMyLecturer/service/EmptySlotService.java
@@ -7,5 +7,5 @@ import java.util.List;
 
 public interface EmptySlotService {
     List<EmptySlotDTO> getAllEmptySlot();
-    EmptySlotDTO creatEmptySlot(int userId, EmptySlot emptySlot);
+    EmptySlotDTO creatEmptySlot(Long userId, EmptySlot emptySlot);
 }
